refactor version to source/common/version
