+++ source/extensions/tracers/datadog/datadog_tracer_impl.cc
@@ -5,11 +5,11 @@
 #include "common/common/enum_to_int.h"
 #include "common/common/fmt.h"
 #include "common/common/utility.h"
-#include "common/common/version.h"
 #include "common/config/utility.h"
 #include "common/http/message_impl.h"
 #include "common/http/utility.h"
 #include "common/tracing/http_tracer_impl.h"
+#include "common/version/version.h"
 
 #include "extensions/tracers/well_known_names.h"
