+++ source/server/admin/BUILD
@@ -220,10 +220,10 @@ envoy_cc_library(
         "//include/envoy/server:admin_interface",
         "//include/envoy/server:instance_interface",
         "//source/common/buffer:buffer_lib",
-        "//source/common/common:version_includes",
         "//source/common/http:codes_lib",
         "//source/common/http:header_map_lib",
         "//source/common/memory:stats_lib",
+        "//source/common/version:version_includes",
         "@envoy_api//envoy/admin/v3:pkg_cc_proto",
     ],
 )
