+++ source/server/config_validation/BUILD
@@ -97,7 +97,6 @@ envoy_cc_library(
         "//source/common/access_log:access_log_manager_lib",
         "//source/common/common:assert_lib",
         "//source/common/common:utility_lib",
-        "//source/common/common:version_lib",
         "//source/common/config:utility_lib",
         "//source/common/grpc:common_lib",
         "//source/common/local_info:local_info_lib",
@@ -106,6 +105,7 @@ envoy_cc_library(
         "//source/common/runtime:runtime_lib",
         "//source/common/stats:stats_lib",
         "//source/common/thread_local:thread_local_lib",
+        "//source/common/version:version_lib",
         "//source/server:configuration_lib",
         "//source/server:server_lib",
         "//source/server/admin:admin_lib",
