+++ test/integration/ads_integration_test.cc
@@ -5,11 +5,11 @@
 #include "envoy/config/route/v3/route.pb.h"
 #include "envoy/grpc/status.h"
 
-#include "common/common/version.h"
 #include "common/config/protobuf_link_hacks.h"
 #include "common/config/version_converter.h"
 #include "common/protobuf/protobuf.h"
 #include "common/protobuf/utility.h"
+#include "common/version/version.h"
 
 #include "test/common/grpc/grpc_client_integration.h"
 #include "test/integration/ads_integration.h"
