+++ tools/testdata/check_format/header_order.cc
@@ -2,7 +2,7 @@
 #include "common/api/api_impl.h"
 #include "common/api/os_sys_calls_impl.h"
 #include "common/common/utility.h"
-#include "common/common/version.h"
+#include "common/version/version.h"
 #include "common/config/resources.h"
 #include "common/config/utility.h"
 #include "common/local_info/local_info_impl.h"
