+++ tools/testdata/check_format/header_order.cc.gold
@@ -18,7 +18,6 @@
 #include "common/api/api_impl.h"
 #include "common/api/os_sys_calls_impl.h"
 #include "common/common/utility.h"
-#include "common/common/version.h"
 #include "common/config/resources.h"
 #include "common/config/utility.h"
 #include "common/local_info/local_info_impl.h"
@@ -30,6 +29,7 @@
 #include "common/singleton/manager_impl.h"
 #include "common/stats/thread_local_store.h"
 #include "common/upstream/cluster_manager_impl.h"
+#include "common/version/version.h"
 
 #include "server/configuration_impl.h"
 #include "server/connection_handler_impl.h"
